Move Extension::Template into file
